change n <<- as.numeric.. to as.integer in glmResp

+++ R/AllClass.R
@@ -485,7 +485,7 @@ glmResp <-
                          ll <- list(...)
                          if (is.null(ll$family)) stop("family must be specified")
                          family <<- ll$family
-                         n <<- if (!is.null(ll$n)) as.numeric(ll$n) else rep.int(1,length(y))
+                         n <<- if (!is.null(ll$n)) as.integer(ll$n) else rep.int(1,length(y))
                          eta <<- numeric(length(y))
                      },
                      aic          = function() {
